+++ .flake8
@@ -1,0 +1,75 @@
+[flake8]
+exclude = .git,sandbox/*,node_modules/
+max-complexity = 10
+max-line-length = 160
+ignore =
+    # Black
+    # E203 - whitespace before ':'. Opposite convention enforced by black
+    E203
+    # E231: missing whitespace after ',', ';', or ':'; for black
+    E231
+    # E501 - line too long. Handled by black, we have longer lines
+    E501
+    # W503 - line break before binary operator, need for black
+    W503
+
+    # McCabe complexity
+    C901,
+
+    # continuation line under-indented for hanging indent
+    E121,
+    # continuation line missing indentation or outdented
+    E122,
+    # closing bracket does not match visual indentation
+    E124,
+    # continuation line with same indent as next logical line
+    E125,
+    # continuation line over-indented for hanging indent
+    E126,
+    # continuation line over-indented for visual indent
+    E127,
+    # continuation line under-indented for visual indent
+    E128,
+    # continuation line unaligned for hanging indent
+    E131,
+    # missing whitespace around operator
+    E225,
+    # missing whitespace around arithmetic operator
+    E226,
+    # multiple spaces after ‘,’
+    E241,
+    # at least two spaces before inline comment
+    E261,
+    # inline comment should start with ‘# ‘
+    E262,
+    # too many leading ‘#’ for block comment
+    E266,
+    # expected 2 blank lines after class or function definition
+    E305,
+    # module level import not at top of file
+    E402,
+    # multiple statements on one line (semicolon)
+    E702,
+    # statement ends with a semicolon
+    E703,
+    # comparison to None should be ‘if cond is None:’
+    E711,
+    # comparison to True should be ‘if cond is True:’ or ‘if cond:’
+    E712,
+    # do not use bare except
+    E722,
+    # do not assign a lambda expression, use a def
+    E731,
+    # mbiguous variable name
+    E741,
+
+    # trailing whitespace
+    W291,
+    # blank line contains whitespace
+    W293,
+    # blank line at end of file
+    W391,
+    # line break occurred before a binary operator
+    W503,
+    # .has_key() is deprecated, use ‘in’
+    W601
