+++ .gitignore
@@ -1,3 +1,14 @@
-.idea/*
+.idea
+venv
 
-*/*.secret
+__pycache__/
+*.py[cod]
+*$py.class
+*.pytest_cache
+
+**/*.secret
+
+.ipynb_checkpoints
+logs
+.cache
+.DS_Store
