--- setups.txt
@@ -1,3 +1,0 @@
-93a1b75a1:Coffee Break! (What if the real cult was the friends we made along the way)
-7a1b57a5b61a1b66a1b31a1:Classic 9 (1 Cop, 1 Medic vs 1 Consigliere, 1 Grandpa...and the villies)
-65a1b75a1b86a1b7a1b32a1b57a4b5a1b31a1:How to Play Marksman (Bulletproof, Cop, Medic, Miller, 4 Villager vs Godfather, Mafia, Marksman)
